requirements.txt
anyio==3.6.2
attrs==22.2.0
base58==2.1.1
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
connexion==2.14.1
Deprecated==1.2.13
ECPy==1.2.5
exceptiongroup==1.1.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Testing==0.8.1
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-secret-manager==2.14.0
googleapis-common-protos==1.57.1
grpc-google-iam-v1==0.12.4
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.18.2
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
packaging==22.0
pluggy==1.0.0
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.7.0
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
swagger-ui-bundle==0.0.9
tomli==2.0.1
types-Deprecated==1.2.9
typing_extensions==4.4.0
urllib3==1.26.13
websockets==11.0
Werkzeug==2.2.2
wrapt==1.14.1
xrpl-py==3.0.0
